+++ CHANGELOG.md
@@ -8,6 +8,7 @@ V0.4.2 - Not released yet
 * Add browser action: Right-CTRL + ◢ + Right - Focus next tab, than reload it 
 * Add action: ◢ + I - Display public and private IP and Wifi MAC address, copy public IP address 
 * Remove unwanted printing character of chording combo in more actions/applications 
+* Add addendum to README: Linux trackball configuration snippets
 
 V0.4.1 - 2021/01/20
 -------------------