+++ CHANGELOG.md
@@ -1,14 +1,15 @@
 xmchord Changelog
 =================
 
-V0.4.2 - Not released yet
--------------------------
+V0.4.2 - 2021/01/29
+-------------------
 * Improve actions that make use of Ctrl+C: Retain previous clipboard content
 * Add browser action: Right-CTRL + ◢ + Left - Focus previous tab, than reload it 
 * Add browser action: Right-CTRL + ◢ + Right - Focus next tab, than reload it 
 * Add action: ◢ + I - Display public and private IP and Wifi MAC address, copy public IP address 
 * Remove unwanted printing character of chording combo in more actions/applications 
 * Add addendum to README: Linux trackball configuration snippets
+* Improve keyboard preference handling (more feedback, allow detect by sub-string)
 
 V0.4.1 - 2021/01/20
 -------------------